Add CI and Rust format/lint codebase

+++ src/jpeg/mod.rs
@@ -410,6 +410,7 @@ fn write_sos(output: &mut Vec<u8>, color_type: ColorType) {
 }
 
 /// Encode the image scan data.
+#[allow(clippy::too_many_arguments)]
 fn encode_scan(
     output: &mut Vec<u8>,
     data: &[u8],
@@ -512,13 +513,13 @@ fn encode_scan(
             }
         }
         (_, Subsampling::S420) => {
-            let padded_width_420 = (width as usize + 15) & !15;
-            let padded_height_420 = (height as usize + 15) & !15;
+            let padded_width_420 = (width + 15) & !15;
+            let padded_height_420 = (height + 15) & !15;
 
             for mcu_y in (0..padded_height_420).step_by(16) {
                 for mcu_x in (0..padded_width_420).step_by(16) {
                     let (y_blocks, cb_block, cr_block) =
-                        extract_mcu_420(data, width as usize, height as usize, mcu_x, mcu_y);
+                        extract_mcu_420(data, width, height, mcu_x, mcu_y);
 
                     for y_block in &y_blocks {
                         let y_dct = dct_2d(y_block);
